requirements.txt
paho-mqtt~=1.6.1
docopt~=0.6.2
google~=3.0.0
google-api-core~=2.4.0
setuptools==60.5.0
adafruit-pca9685